Restructure editor files
