+++ src/editor/input/SingleObjectEditor.tsx
